Refactor node.go into multiple files
- Refactored node.go into multiple files because it was too big and hard to manage.
- Also wrote some tests for a couple of the refactored files.
