+++ pkg/assembler/contains_edge.go
@@ -1,0 +1,43 @@
+//
+// Copyright 2023 The GUAC Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package assembler
+
+// ContainsEdge Contains is an edge that represents the fact that an
+// `PackageNode` contains a `ArtifactNode`
+type ContainsEdge struct {
+	PackageNode       PackageNode
+	ContainedArtifact ArtifactNode
+}
+
+func (e ContainsEdge) Type() string {
+	return "Contains"
+}
+
+func (e ContainsEdge) Nodes() (v, u GuacNode) {
+	return e.PackageNode, e.ContainedArtifact
+}
+
+func (e ContainsEdge) Properties() map[string]interface{} {
+	return map[string]interface{}{}
+}
+
+func (e ContainsEdge) PropertyNames() []string {
+	return []string{}
+}
+
+func (e ContainsEdge) IdentifiablePropertyNames() []string {
+	return []string{}
+}
