+++ pkg/assembler/depends_on_edge.go
@@ -1,0 +1,68 @@
+//
+// Copyright 2023 The GUAC Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package assembler
+
+// DependsOnEdge is an edge that represents the fact that an
+// `ArtifactNode/PackageNode` depends on another `ArtifactNode/PackageNode`
+// Only one of each side of the edge should be defined.
+type DependsOnEdge struct {
+	ArtifactNode       ArtifactNode
+	PackageNode        PackageNode
+	ArtifactDependency ArtifactNode
+	PackageDependency  PackageNode
+}
+
+func (e DependsOnEdge) Type() string {
+	return "DependsOn"
+}
+
+func (e DependsOnEdge) Nodes() (v, u GuacNode) {
+	vA, vP := isDefined(e.ArtifactNode), isDefined(e.PackageNode)
+	uA, uP := isDefined(e.ArtifactDependency), isDefined(e.PackageDependency)
+	if vA == vP {
+		panic("only one of package and artifact node defined for DependsOn relationship")
+	}
+
+	if uA == uP {
+		panic("only one of package and artifact dependency node defined for DependsOn relationship")
+	}
+
+	if vA {
+		v = e.ArtifactNode
+	} else {
+		v = e.PackageNode
+	}
+
+	if uA {
+		u = e.ArtifactDependency
+	} else {
+		u = e.PackageDependency
+	}
+
+	return v, u
+}
+
+func (e DependsOnEdge) Properties() map[string]interface{} {
+	return map[string]interface{}{}
+}
+
+func (e DependsOnEdge) PropertyNames() []string {
+	return []string{}
+}
+
+func (e DependsOnEdge) IdentifiablePropertyNames() []string {
+	return []string{}
+}
