+++ pkg/assembler/vulnerability_node.go
@@ -1,0 +1,44 @@
+//
+// Copyright 2023 The GUAC Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package assembler
+
+// VulnerabilityNode is a node that represents a vulnerability associated with the certifier attestation
+type VulnerabilityNode struct {
+	ID       string
+	NodeData objectMetadata
+}
+
+func (vn VulnerabilityNode) Type() string {
+	return "Vulnerability"
+}
+
+func (vn VulnerabilityNode) Properties() map[string]interface{} {
+	properties := make(map[string]interface{})
+	properties["id"] = vn.ID
+	vn.NodeData.addProperties(properties)
+	return properties
+}
+
+func (vn VulnerabilityNode) PropertyNames() []string {
+	fields := []string{"id"}
+	fields = append(fields, vn.NodeData.getProperties()...)
+	return fields
+}
+
+func (vn VulnerabilityNode) IdentifiablePropertyNames() []string {
+	// Based on the ID of the vulnerability, more information can be obtained but not stored in the graph DB
+	return []string{"id"}
+}
